Add files via upload
The code and dataset are focused on the task of recognizing handwritten alphabets. In this context, the dataset includes defined labels for each alphabet, and the code employs a Convolutional Neural Network (CNN) model to perform the recognition. Additionally, the code allows for user input at the end, enabling the assessment of the model's accuracy based on custom inputs.